- H2 Embeded, Exception handler

+++ src/main/java/com/rest/api/repo/UserJpaRepo.java
@@ -4,5 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.rest.api.entity.User;
 
-public interface UserJpaRepo extends JpaRepository<User, Integer> {
+public interface UserJpaRepo extends JpaRepository<User, Long> {
 }
